refactor(shim): PR-9.2 — 删 browse shim（proof-of-concept）
shim 删除安全策略验证：
1. grep 全部 callers（含 internal `from .. import X` 嵌套层）
2. 每 caller 显式改 canonical path（不用 __getattr__ lazy 回退，fail loud + early）
3. 删 shim 文件
4. 跑 full sweep 确认 baseline 不变

  studio/browse.py             删（10 行 shim → 0）
  studio/api/routers/browse.py from ... import browse, datasets →
                               from ...services.dataset import browse, scan as datasets
                               + from ...paths import REPO_ROOT → infrastructure.paths
  tests/test_browse.py         from studio import browse, server →
                               拆 from studio import server / from studio.services.dataset import browse

回归：1568 passed broader sweep（== baseline，零 regression）。

剩余 45 shims 同模式（per-shim commit + full sweep gate）：
- 11 个 top-level → services shims (curation/datasets/preprocess/presets_io/
  project_jobs/projects/queue_io/task_snapshot/thumb_cache/versions/versions_phase)
- 8 个 top-level → infrastructure shims (paths/db/secrets/event_bus/log_tail/
  argparse_bridge/llm_presets/migrations)
- 29 个 services/ shims
- 2 个永久保留 (server.py 给 SPA mount + 入口；db/paths/secrets 也可考虑删但
  test fixture monkeypatch 量大风险高，暂留)
